deep_learning/brain_tumor/requirements.txt
numpy==1.26.4
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0

tensorflow==2.15.0

torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0

xgboost>=2.0.0
joblib>=1.3.0

tqdm>=4.66.0